improved state machine
