+++ package.json
@@ -10,6 +10,7 @@
     "start": "next start"
   },
   "dependencies": {
+    "@xstate/react": "^0.8.1",
     "@zeit/next-sass": "^1.0.1",
     "bootstrap": "^4.4.1",
     "express": "^4.17.1",
@@ -18,10 +19,13 @@
     "next-compose-plugins": "^2.2.0",
     "node-sass": "^4.13.0",
     "react": "16.12.0",
-    "react-dom": "16.12.0"
+    "react-dom": "16.12.0",
+    "xstate": "^4.8.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^5.0.0",
+    "eslint": "^6.8.0",
+    "eslint-plugin-react": "^7.19.0",
     "prettier": "2.0.2"
   }
 }
