+++ utils/audioPlayer.js
@@ -18,8 +18,8 @@ class AudioPlayer {
       howler.stop();
       howler.unload();
       subscribersStopped
-        .filter((s) => s.id !== book.id)
-        .forEach((s) => s.callback());
+        .filter(s => s.id !== book.id)
+        .forEach(s => s.callback());
     }
 
     this._createHowlerObject(book, events, trackIndex);
@@ -33,7 +33,7 @@ class AudioPlayer {
         JSON.stringify({
           book,
           trackIndex,
-          elapsedTime,
+          elapsedTime
         })
       );
   }
@@ -50,14 +50,15 @@ class AudioPlayer {
           1000
         );
         subscribersStarted
-          .filter((s) => s.id === book.id)
-          .forEach((s) => s.callback(trackIndex));
+          .filter(s => s.id === book.id)
+          .forEach(s => s.callback(trackIndex));
+        if (events.onPlay) events.onPlay();
       },
       onpause: () => {
         clearInterval(progressInterval);
         subscribersPaused
-          .filter((s) => s.id === book.id)
-          .forEach((s) => s.callback());
+          .filter(s => s.id === book.id)
+          .forEach(s => s.callback());
       },
       onload: () => {
         if (events.onLoad) events.onLoad();
@@ -72,7 +73,7 @@ class AudioPlayer {
           this._createHowlerObject(book, events, trackIndex + 1);
           this.play();
         }
-      },
+      }
     });
   }
